JPA integration succesfully added

+++ demo/src/main/java/com/example/demo/service/ClienteService.java
@@ -2,17 +2,22 @@ package com.example.demo.service;
 
 import java.util.Collection;
 
+import org.springframework.stereotype.Service;
+
 import com.example.demo.model.Cliente;
+import com.example.demo.model.Mascota;
 
 public interface ClienteService {
     
-    public Cliente SearchById(int id);
+    public Cliente SearchById(Long id);
 
     public Collection<Cliente> SearchAll();
 
-    public void deleteById(int id);
+    public void deleteById(Long id);
 
     public void update(Cliente cliente);
 
     public void add(Cliente cliente);
+
+    
 }
